pyproject.toml
[project]
name = "stock_research_assistant"
version = "0.1.0"
description = "stock_research_assistant using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "finnhub-python>=2.4.23",
    "pydantic>=2.11.3",
    "setuptools>=80.0.0",
]

[project.scripts]
stock_research_assistant = "stock_research_assistant.main:run"
run_crew = "stock_research_assistant.main:run"
train = "stock_research_assistant.main:train"
replay = "stock_research_assistant.main:replay"
test = "stock_research_assistant.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
